% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shapes.R
\name{circle.area}
\alias{circle.area}
\title{Calculate the area of a circle}
\usage{
circle.area(radius = NULL)
}
\arguments{
\item{radius}{A number}
}
\value{
A number
}
\description{
Calculate the area of a circle
}
\examples{
circle.area(5)
circle.area(22.5)
}
