% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shapes.R
\name{circle.perimeter}
\alias{circle.perimeter}
\title{Calculate the perimeter of a circle}
\usage{
circle.perimeter(radius = NULL)
}
\arguments{
\item{radius}{A number}
}
\value{
A number
}
\description{
Calculate the perimeter of a circle
}
\examples{
circle.perimeter(5)
circle.perimeter(22.5)
}
