% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shapes.R
\name{quadrilateral.area}
\alias{quadrilateral.area}
\title{Calculate the area of a quadrilateral}
\usage{
quadrilateral.area(length = NULL, width = NULL)
}
\arguments{
\item{length}{A number}

\item{width}{A number}
}
\value{
A number
}
\description{
Calculate the area of a quadrilateral
}
\examples{
quadrilateral.area(5, 10)
quadrilateral.area(22.5. 9.9)
}
