% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shapes.R
\name{quadrilateral.perimeter}
\alias{quadrilateral.perimeter}
\title{Calculate the perimeter of a quadrilateral}
\usage{
quadrilateral.perimeter(length = NULL, width = NULL)
}
\arguments{
\item{length}{A number}

\item{width}{A number}
}
\value{
A number
}
\description{
Calculate the perimeter of a quadrilateral
}
\examples{
quadrilateral.perimeter(5, 10)
quadrilateral.perimeter(22.5. 9.9)
}
