% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shapes.R
\name{regular.pentagon.area}
\alias{regular.pentagon.area}
\title{Calculate the area of a regular pentagon}
\usage{
regular.pentagon.area(side = NULL)
}
\arguments{
\item{side}{A number}
}
\value{
A number
}
\description{
Calculate the area of a regular pentagon
}
\examples{
regular.pentagon.area(5)
regular.pentagon.area(22.5)
}
