% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shapes.R
\name{regular.pentagon.perimeter}
\alias{regular.pentagon.perimeter}
\title{Calculate the perimeter of a regular pentagon}
\usage{
regular.pentagon.perimeter(side = NULL)
}
\arguments{
\item{side}{A number}
}
\value{
A number
}
\description{
Calculate the perimeter of a regular pentagon
}
\examples{
regular.pentagon.perimeter(5)
regular.pentagon.perimeter(22.5)
}
