% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shapes.R
\name{square.area}
\alias{square.area}
\title{Calculate the area of a square}
\usage{
square.area(side = NULL)
}
\arguments{
\item{side}{A number}
}
\value{
A number
}
\description{
Calculate the area of a square
}
\examples{
square.area(5)
square.area(22.5)
}
