% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shapes.R
\name{square.perimeter}
\alias{square.perimeter}
\title{Calculate the perimeter of a square}
\usage{
square.perimeter(side = NULL)
}
\arguments{
\item{radius}{A number}
}
\value{
A number
}
\description{
Calculate the perimeter of a square
}
\examples{
square.perimeter(5)
square.perimeter(22.5)
}
