% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shapes.R
\name{triangle.area.heron.formula}
\alias{triangle.area.heron.formula}
\title{Calculate the area of a triangle where all 3 sides are known}
\usage{
triangle.area.heron.formula(side.A = NULL, side.B = NULL, side.C = NULL)
}
\arguments{
\item{side.A}{A number}

\item{side.B}{A number}

\item{side.C}{A number}
}
\value{
A number
}
\description{
Calculate the area of a triangle where all 3 sides are known
}
\examples{
triangle.area.heron.formula(5, 5, 5)
triangle.area.heron.formula(22.5, 14.6, 11.9)
}
