% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shapes.R
\name{triangle.perimeter}
\alias{triangle.perimeter}
\title{Calculate the perimeter of a triangle where all 3 sides are known}
\usage{
triangle.perimeter(side.A = NULL, side.B = NULL, side.C = NULL)
}
\arguments{
\item{side.A}{A number}

\item{side.B}{A number}

\item{side.C}{A number}
}
\value{
A number
}
\description{
Calculate the perimeter of a triangle where all 3 sides are known
}
\examples{
triangle.perimeter(5, 5, 5)
triangle.perimeter(22.5, 14.6, 11.9)
}
