% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shapes.R
\name{triangle.perimeter.SAS.formula}
\alias{triangle.perimeter.SAS.formula}
\title{Calculate the perimeter of a triangle where 2 sides and 1 angle are known}
\usage{
triangle.perimeter.SAS.formula(side.A = NULL, side.B = NULL, angle.C = NULL)
}
\arguments{
\item{side.A}{A number}

\item{side.B}{A number}

\item{angle.C}{A number, angle measured in radians}
}
\value{
A number
}
\description{
Calculate the perimeter of a triangle where 2 sides and 1 angle are known
}
\examples{
triangle.perimeter.SAS.formula(5, 5, 5)
triangle.perimeter.SAS.formula(22.5, 14.6, )
}
